Moved files to HandwritingLibrary
