+++ Handwriting/HandwritingLibrary/Properties/Resources.resx
